remove inner classes
