+++ infrastructure/yaml/src/main/java/tech/pegasys/teku/data/yaml/UInt64Deserializer.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright Consensys Software Inc., 2024
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.data.yaml;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import java.io.IOException;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+
+public class UInt64Deserializer extends JsonDeserializer<UInt64> {
+
+  @Override
+  public UInt64 deserialize(final JsonParser p, final DeserializationContext ctxt)
+      throws IOException {
+    return UInt64.valueOf(p.getValueAsString());
+  }
+}
