+++ infrastructure/yaml/src/main/java/tech/pegasys/teku/data/yaml/UInt64Serializer.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright Consensys Software Inc., 2024
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.data.yaml;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+import java.io.IOException;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+
+public class UInt64Serializer extends JsonSerializer<UInt64> {
+
+  @Override
+  public void serialize(
+      final UInt64 value, final JsonGenerator gen, final SerializerProvider serializers)
+      throws IOException {
+    gen.writeNumber(value.bigIntegerValue());
+  }
+}
