+++ infrastructure/yaml/src/main/java/tech/pegasys/teku/data/yaml/YamlProvider.java
@@ -14,26 +14,18 @@
 package tech.pegasys.teku.data.yaml;
 
 import com.fasterxml.jackson.core.JsonGenerationException;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.Module;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.StringWriter;
 import java.io.UncheckedIOException;
-import java.util.Locale;
 import java.util.stream.Stream;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
@@ -43,22 +35,19 @@ public class YamlProvider {
   private final ObjectMapper objectMapper;
 
   public YamlProvider(final Module... modules) {
-    this.objectMapper = new ObjectMapper(new YAMLFactory());
+    objectMapper = new ObjectMapper(new YAMLFactory());
     addTekuMappers();
     Stream.of(modules).forEach(objectMapper::registerModule);
   }
 
   private void addTekuMappers() {
-    SimpleModule module = new SimpleModule("TekuYaml", new Version(1, 0, 0, null, null, null));
+    final SimpleModule module =
+        new SimpleModule("TekuYaml", new Version(1, 0, 0, null, null, null));
     module.addDeserializer(UInt64.class, new UInt64Deserializer());
     module.addSerializer(UInt64.class, new UInt64Serializer());
     module.addDeserializer(Bytes32.class, new Bytes32Deserializer());
     module.addSerializer(Bytes.class, new BytesSerializer());
     objectMapper.registerModule(module).writer(new DefaultPrettyPrinter());
-  }
-
-  public <T> T read(final InputStream data, final Class<T> clazz) throws IOException {
-    return objectMapper.readValue(data, clazz);
   }
 
   public <T> T read(final Bytes data, final Class<T> clazz) throws IOException {
@@ -93,41 +82,5 @@ public class YamlProvider {
 
   public ObjectMapper getObjectMapper() {
     return objectMapper;
-  }
-
-  public static class UInt64Deserializer extends JsonDeserializer<UInt64> {
-
-    @Override
-    public UInt64 deserialize(final JsonParser p, final DeserializationContext ctxt)
-        throws IOException {
-      return UInt64.valueOf(p.getValueAsString());
-    }
-  }
-
-  public static class UInt64Serializer extends JsonSerializer<UInt64> {
-    @Override
-    public void serialize(
-        final UInt64 value, final JsonGenerator gen, final SerializerProvider serializers)
-        throws IOException {
-      gen.writeNumber(value.bigIntegerValue());
-    }
-  }
-
-  public static class Bytes32Deserializer extends JsonDeserializer<Bytes32> {
-
-    @Override
-    public Bytes32 deserialize(final JsonParser p, final DeserializationContext ctxt)
-        throws IOException {
-      return Bytes32.fromHexString(p.getValueAsString());
-    }
-  }
-
-  public static class BytesSerializer extends JsonSerializer<Bytes> {
-    @Override
-    public void serialize(
-        final Bytes value, final JsonGenerator gen, final SerializerProvider provider)
-        throws IOException {
-      gen.writeString(value.toHexString().toLowerCase(Locale.ROOT));
-    }
   }
 }
